requirements.txt
asn1crypto==0.22.0
boto3==1.4.6
botocore==1.6.8
cffi==1.10.0
cryptography==2.0.3
Django==1.11.4
django-cors-headers==2.1.0
django-encrypted-fields==1.1.2
django-sslserver==0.20
djangorestframework==3.6.4
docutils==0.14
enum34==1.1.6
futures==3.1.1
idna==2.6
ipaddress==1.0.18
jmespath==0.9.3
psycopg2==2.7.3
pyasn1==0.3.2
pycparser==2.18
pycrypto==2.6.1
pyotp==2.2.6
python-dateutil==2.6.1
python-keyczar==0.716
python-u2flib-server==5.0.0
pytz==2017.2
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
